feat: support manually retrying to initiate a request
