+++ package.json
@@ -158,7 +158,6 @@
     "react-markdown": "^8.0.7",
     "react-syntax-highlighter": "^15.5.0",
     "react-virtualized": "^9.22.5",
-    "remark-gfm": "^4.0.0",
     "tslog": "^4.9.2"
   },
   "repository": {
@@ -187,7 +186,9 @@
     },
     "plugins": {
       "@release-it/conventional-changelog": {
-        "preset": "angular",
+        "preset": {
+          "name": "angular"
+        },
         "infile": "CHANGELOG.md"
       }
     }
