+++ src/isomorphic/consts.ts
@@ -18,5 +18,5 @@ export enum MSG_LEVEL {
 }
 
 export const DEFAULT_RESPONSE_MESSAGE = `Failed to get answers, pls retry or check your cookies.
-You can obtain your cookie by following the instructions provided in this [link](https://github.com/JimmyDaddy/vscode-bard#how-to-get-cookies).
+You can get your cookie by following the instructions provided in this [link](https://github.com/JimmyDaddy/vscode-bard#how-to-get-cookies).
 `;
