+++ src/view/App.module.less
@@ -24,71 +24,6 @@
   } 
 }
 
-.conversationContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-  cursor: default;
-}
-
-.contentContainer {
-  text-align: left;
-  border-bottom: #bdbdbd57 solid 1px;
-  width: 100%;
-  padding: 16px 0px;
-  cursor: pointer;
-}
-
-.person {
-  display: flex;
-  align-items: center;
-  text-align: left;
-  width: 100%;
-
-  .deleteContainer {
-    flex: 1;
-    text-align: right;
-
-    &.hidden {
-      display: none;
-    }
-  }
-}
-
-.name {
-  margin-left: 8px;
-  font-weight: bold;
-}
-
-.content {
-  margin-top: 16px;
-}
-
-.responseContainer {
-  margin-left: 0px;
-}
-
-.promptSelectContainer {
-  flex: 1;
-  text-align: right;
-}
-
-.promptSelectRoot {
-  svg {
-    color: #bdbdbd57;
-  }
-}
-
-.promptSelect {
-  text-align: right;
-  font-size: small !important;
-  color: #bdbdbd57 !important;
-}
-
 .loading {
   font-size: 24px;
 }
