+++ src/view/App.tsx
@@ -1,148 +1,19 @@
 
 import React, { useState, useEffect, useRef, useCallback } from 'react';
-import ReactMarkdown from 'react-markdown';
 import Box from '@mui/material/Box';
 import IconButton from '@mui/material/IconButton';
-import Avatar from '@mui/material/Avatar';
 import Input from '@mui/material/InputBase';
-import Select from '@mui/material/NativeSelect';
 import SendIcon from '@mui/icons-material/Send';
 import CircularProgress from '@mui/material/CircularProgress';
-import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import AutoAwesomeIcon from '@mui/icons-material/AutoAwesomeSharp';
-import DeleteIcon from '@mui/icons-material/Delete';
 import List from 'react-virtualized/dist/commonjs/List';
 import AutoSizer from 'react-virtualized/dist/commonjs/AutoSizer';
 import CellMeasurer, { CellMeasurerCache } from 'react-virtualized/dist/commonjs/CellMeasurer';
-import { grey } from '@mui/material/colors';
 
 import { uid, debounce } from '../isomorphic/utils';
 import { MSG } from '../isomorphic/consts';
-import CodeBlock from './CodeBlock';
+import Conversation from './Conversation';
 
 import styles from './App.module.less';
-
-function Row(props: {
-  index: number,
-  style: any,
-  data: BardMessage,
-  measure: any,
-}) {
-  const { index, data: curData } = props;
-  const [response, setResponse] = useState<BardResponse | undefined>(curData?.responses?.[0]);
-  const [deleteHover, setDeleteHover] = useState(false);
-  if (!curData) {
-    return null;
-  }
-
-  useEffect(() => {
-    props.measure?.();
-  }, []);
-
-  useEffect(() => {
-    setResponse(curData.responses?.[0]);
-  }, [curData]);
-
-  useEffect(() => {
-    props.measure?.();
-  }, [response]);
-
-  const handleCHange = (event: any) => {
-    const value = event?.target?.value as string;
-    const response = curData?.responses?.find((response: BardResponse) => response.prompt === value);
-    if (response) {
-      setResponse(response);
-    }
-  };
-
-  function deleteMsg() {
-    vscode.postMessage({ 
-      type: MSG.deleteMessage, 
-      message: {
-        uid: curData.uid,
-      },
-    });
-  }
-
-  return (
-    <div key={index} className={styles.conversationContainer} style={props.style}>
-      <div
-        className={styles.contentContainer}
-        onMouseOver={() => {
-          setDeleteHover(true);
-        }}
-        onMouseLeave={() => {
-          setDeleteHover(false);
-        }}
-      >
-        <div
-          className={styles.person}
-        >
-          <Avatar sx={{ width: 30, height: 30 }}>
-            <AccountCircleIcon sx={{ fontSize: 32 }} />
-          </Avatar>
-          <span className={styles.name}>You</span>
-          <div
-            className={`${styles.deleteContainer} ${deleteHover? '' : styles.hidden}`} 
-          >
-            <IconButton aria-label="delete" size="small" onClick={deleteMsg}>
-              <DeleteIcon fontSize="inherit" sx={{ color: grey['A700'] }} />
-            </IconButton>
-          </div>
-        </div>
-        <div className={styles.content}>
-          <ReactMarkdown>
-            {curData.ask || ''}
-          </ReactMarkdown>
-        </div>
-      </div>
-      <div className={styles.contentContainer}>
-        <div className={styles.person}>
-          <Avatar sx={{ width: 30, height: 30 }}>
-            <AutoAwesomeIcon />
-          </Avatar>
-          <span className={styles.name}>Bard</span>
-          {
-            curData.responses?.length !== undefined && curData.responses?.length > 1 && (
-              <div className={styles.promptSelectContainer}>
-                <Select
-                  id={`prompt-selector-${index}`}
-                  value={response?.prompt}
-                  onChange={handleCHange}
-                  variant='standard'
-                  className={styles.promptSelectRoot}
-                  classes={{
-                    standard: styles.promptSelect,
-                  }}
-                >
-                  {curData.responses?.map((response: BardResponse, index: number) => (
-                    <option
-                      key={`${response.prompt}${index}response`}
-                      value={response.prompt}
-                    >
-                      {response.prompt}
-                    </option>
-                  ))}
-                </Select>
-            </div>
-            )
-          }
-        </div>
-        <div className={styles.responseContainer}>
-          <ReactMarkdown
-            key={`${index}response`}
-            children={`${response?.response || ''}`}
-            components={{
-              code(props) {
-                return <CodeBlock {...props }/>;
-              }
-            }}
-          />
-        </div>
-      </div>
-    </div>
-  );
-};
 
 function App() {
   const [data, setData] = useState<BardMessage[]>([]);
@@ -195,13 +66,17 @@ function App() {
       inputRef.current.value = '';
     }
     vscode.postMessage({ 
-      type: 'sendMessage', 
+      type: MSG.sendMessage, 
       message: {
         prompt: message,
         uid: msgUid,
       },
     });
   }
+
+  const retry = () => {
+    setLoading(true);
+  };
 
   const renderRow = ({ index, key, style, parent }: any) => {
     return (
@@ -213,21 +88,30 @@ function App() {
         parent={parent}
       >
         {({ measure }) => (
-          <Row
+          <Conversation
             index={index}
             style={style}
             data={data[index]}
             measure={measure}
+            onRetry={retry}
           />
         )}
       </CellMeasurer>
     );
   };
 
-  window.addEventListener('message', (event) => {
+  const messageHandler = useCallback((event: any) => {
     const msgData = event.data;
     if (msgData.type === MSG.showResponse) {
-      setData(data.filter(d => !d.isTemp).concat(JSON.parse(msgData.data)));
+      let newNonTempData = data.filter(d => !d.isTemp);
+      const responseData = JSON.parse(msgData.data);
+      const existIndex = data.findIndex((v) => v.uid === responseData.uid);
+      if (existIndex === -1) {
+        newNonTempData = newNonTempData.concat([responseData]);
+      } else {
+        newNonTempData[existIndex] = responseData;
+      }
+      setData(newNonTempData);
       setLoading(false);
       setMessage('');  
       return;
@@ -243,7 +127,14 @@ function App() {
       setData(d.conversations || []);
       setInputHistory(d.promptHistory || []);
     }
-  });
+  }, [data]);
+
+  useEffect(() => {
+    window.addEventListener('message', messageHandler);
+    return () => {
+      window.removeEventListener('message', messageHandler);
+    };
+  }, [data]);
 
   const handleKeyDown = (event: any) => {
     if (event.keyCode === 13 && !event.shiftKey) {
@@ -324,4 +215,3 @@ function App() {
   );
 }
 export default App;
-
