+++ src/view/CodeBlock.module.less
@@ -22,4 +22,3 @@
     }
   }
 }
-
